+++ cli/src/cli_util.rs
@@ -3867,7 +3867,7 @@ impl<'a> CliRunner<'a> {
                      accessed?",
                 )
             })?;
-        let mut config_env = ConfigEnv::from_environment(ui);
+        let mut config_env = ConfigEnv::from_environment();
         let mut last_config_migration_descriptions = Vec::new();
         let mut migrate_config = |config: &mut StackedConfig| -> Result<(), CommandError> {
             last_config_migration_descriptions =