+++ cli/src/complete.rs
@@ -1058,8 +1058,7 @@ fn get_jj_command() -> Result<(JjBuilder, UserSettings), CommandError> {
     let cwd = std::env::current_dir()
         .and_then(dunce::canonicalize)
         .map_err(user_error)?;
-    // No config migration for completion. Simply ignore deprecated variables.
-    let mut config_env = ConfigEnv::from_environment(&ui);
+    let mut config_env = ConfigEnv::from_environment();
     let maybe_cwd_workspace_loader = DefaultWorkspaceLoaderFactory.create(find_workspace_dir(&cwd));
     let _ = config_env.reload_user_config(&mut raw_config);
     if let Ok(loader) = &maybe_cwd_workspace_loader {