+++ cli/src/config.rs
@@ -45,7 +45,6 @@ use crate::command_error::CommandError;
 use crate::command_error::config_error;
 use crate::command_error::config_error_with_message;
 use crate::text_util;
-use crate::ui::Ui;
 
 // TODO(#879): Consider generating entire schema dynamically vs. static file.
 pub const CONFIG_SCHEMA: &str = include_str!("config-schema.json");
@@ -267,14 +266,12 @@ fn create_dir_all(path: &Path) -> std::io::Result<()> {
 #[derive(Clone, Default, Debug)]
 struct UnresolvedConfigEnv {
     config_dir: Option<PathBuf>,
-    // TODO: remove after jj 0.35
-    macos_legacy_config_dir: Option<PathBuf>,
     home_dir: Option<PathBuf>,
     jj_config: Option<String>,
 }
 
 impl UnresolvedConfigEnv {
-    fn resolve(self, ui: &Ui) -> Vec<ConfigPath> {
+    fn resolve(self) -> Vec<ConfigPath> {
         if let Some(paths) = self.jj_config {
             return split_paths(&paths)
                 .filter(|path| !path.as_os_str().is_empty())
@@ -297,21 +294,9 @@ impl UnresolvedConfigEnv {
             config_dir.push("conf.d");
             ConfigPath::new(config_dir)
         });
-        let legacy_platform_config_path =
-            self.macos_legacy_config_dir.clone().map(|mut config_dir| {
-                config_dir.push("jj");
-                config_dir.push("config.toml");
-                ConfigPath::new(config_dir)
-            });
-        let legacy_platform_config_dir = self.macos_legacy_config_dir.map(|mut config_dir| {
-            config_dir.push("jj");
-            config_dir.push("conf.d");
-            ConfigPath::new(config_dir)
-        });
 
         if let Some(path) = home_config_path
-            && (path.exists()
-                || (platform_config_path.is_none() && legacy_platform_config_path.is_none()))
+            && (path.exists() || platform_config_path.is_none())
         {
             paths.push(path);
         }
@@ -328,42 +313,7 @@ impl UnresolvedConfigEnv {
             paths.push(path);
         }
 
-        if let Some(path) = legacy_platform_config_path
-            && path.exists()
-        {
-            Self::warn_for_deprecated_path(
-                ui,
-                path.as_path(),
-                "~/Library/Application Support/jj",
-                "~/.config/jj",
-            );
-            paths.push(path);
-        }
-        if let Some(path) = legacy_platform_config_dir
-            && path.exists()
-        {
-            Self::warn_for_deprecated_path(
-                ui,
-                path.as_path(),
-                "~/Library/Application Support/jj",
-                "~/.config/jj",
-            );
-            paths.push(path);
-        }
-
         paths
-    }
-
-    fn warn_for_deprecated_path(ui: &Ui, path: &Path, old: &str, new: &str) {
-        let _ = indoc::writedoc!(
-            ui.warning_default(),
-            r"
-            Deprecated configuration file `{}`.
-            Configuration files in `{old}` are deprecated, and support will be removed in a future release.
-            Instead, move your configuration files to `{new}`.
-            ",
-            path.display(),
-        );
     }
 }
 
@@ -381,28 +331,10 @@ pub struct ConfigEnv {
 
 impl ConfigEnv {
     /// Initializes configuration loader based on environment variables.
-    pub fn from_environment(ui: &Ui) -> Self {
+    pub fn from_environment() -> Self {
         let config_dir = etcetera::choose_base_strategy()
             .ok()
             .map(|s| s.config_dir());
-
-        // older versions of jj used a more "GUI" config option,
-        // which is not designed for user-editable configuration of CLI utilities.
-        let macos_legacy_config_dir = if cfg!(target_os = "macos") {
-            etcetera::base_strategy::choose_native_strategy()
-                .ok()
-                .map(|s| {
-                    // note that etcetera calls Library/Application Support the "data dir",
-                    // Library/Preferences is supposed to be exclusively plists
-                    s.data_dir()
-                })
-                .filter(|data_dir| {
-                    // User might've purposefully set their config dir to the deprecated one
-                    Some(data_dir) != config_dir.as_ref()
-                })
-        } else {
-            None
-        };
 
         // Canonicalize home as we do canonicalize cwd in CliRunner. $HOME might
         // point to symlink.
@@ -412,7 +344,6 @@ impl ConfigEnv {
 
         let env = UnresolvedConfigEnv {
             config_dir,
-            macos_legacy_config_dir,
             home_dir: home_dir.clone(),
             jj_config: env::var("JJ_CONFIG").ok(),
         };
@@ -420,7 +351,7 @@ impl ConfigEnv {
             home_dir,
             repo_path: None,
             workspace_path: None,
-            user_config_paths: env.resolve(ui),
+            user_config_paths: env.resolve(),
             repo_config_path: None,
             workspace_config_path: None,
             command: None,
@@ -1727,51 +1658,6 @@ mod tests {
         }
     }
 
-    fn config_path_macos_legacy_exists() -> TestCase {
-        TestCase {
-            files: &["macos-legacy/jj/config.toml"],
-            env: UnresolvedConfigEnv {
-                home_dir: Some("home".into()),
-                config_dir: Some("config".into()),
-                macos_legacy_config_dir: Some("macos-legacy".into()),
-                ..Default::default()
-            },
-            wants: vec![
-                Want::new("config/jj/config.toml"),
-                Want::existing("macos-legacy/jj/config.toml"),
-            ],
-        }
-    }
-
-    fn config_path_macos_legacy_both_exist() -> TestCase {
-        TestCase {
-            files: &["macos-legacy/jj/config.toml", "config/jj/config.toml"],
-            env: UnresolvedConfigEnv {
-                home_dir: Some("home".into()),
-                config_dir: Some("config".into()),
-                macos_legacy_config_dir: Some("macos-legacy".into()),
-                ..Default::default()
-            },
-            wants: vec![
-                Want::existing("config/jj/config.toml"),
-                Want::existing("macos-legacy/jj/config.toml"),
-            ],
-        }
-    }
-
-    fn config_path_macos_legacy_new() -> TestCase {
-        TestCase {
-            files: &[],
-            env: UnresolvedConfigEnv {
-                home_dir: Some("home".into()),
-                config_dir: Some("config".into()),
-                macos_legacy_config_dir: Some("macos-legacy".into()),
-                ..Default::default()
-            },
-            wants: vec![Want::new("config/jj/config.toml")],
-        }
-    }
-
     #[test_case(config_path_home_existing())]
     #[test_case(config_path_home_new())]
     #[test_case(config_path_home_existing_platform_new())]
@@ -1790,9 +1676,6 @@ mod tests {
     #[test_case(config_path_platform_existing_conf_dir_existing())]
     #[test_case(config_path_all_existing())]
     #[test_case(config_path_none())]
-    #[test_case(config_path_macos_legacy_exists())]
-    #[test_case(config_path_macos_legacy_both_exist())]
-    #[test_case(config_path_macos_legacy_new())]
     fn test_config_path(case: TestCase) {
         let tmp = setup_config_fs(case.files);
         let env = resolve_config_env(&case.env, tmp.path());
@@ -1832,7 +1715,6 @@ mod tests {
         let home_dir = env.home_dir.as_ref().map(|p| root.join(p));
         let env = UnresolvedConfigEnv {
             config_dir: env.config_dir.as_ref().map(|p| root.join(p)),
-            macos_legacy_config_dir: env.macos_legacy_config_dir.as_ref().map(|p| root.join(p)),
             home_dir: home_dir.clone(),
             jj_config: env.jj_config.as_ref().map(|p| {
                 join_paths(split_paths(p).map(|p| {
@@ -1850,7 +1732,7 @@ mod tests {
             home_dir,
             repo_path: None,
             workspace_path: None,
-            user_config_paths: env.resolve(&Ui::null()),
+            user_config_paths: env.resolve(),
             repo_config_path: None,
             workspace_config_path: None,
             command: None,